+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
+## 2023-10-09 - 0.0.2
+
+-   Add command alias
+-   Change description
+
 ## 2023-10-06 - 0.0.1
 
 -   Add boilerplate
 -   Add first test command
+-   Add all commands from legacy bot